+++ src/main/kotlin/org/jetbrains/research/testspark/services/TestStorageProcessingService.kt
@@ -10,9 +10,11 @@ import com.intellij.openapi.roots.CompilerModuleExtension
 import com.intellij.openapi.roots.ModuleRootManager
 import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.util.io.FileUtilRt
+import org.jetbrains.research.testspark.DataFilesUtil
 import org.jetbrains.research.testspark.data.TestCase
 import org.jetbrains.research.testspark.editor.Workspace
 import org.jetbrains.research.testspark.tools.getBuildPath
+import org.jetbrains.research.testspark.tools.llm.test.TestCaseGeneratedByLLM
 import java.io.File
 import java.util.UUID
 import kotlin.collections.ArrayList
@@ -20,7 +22,7 @@ import kotlin.io.path.Path
 import kotlin.io.path.createDirectories
 
 @Service(Service.Level.PROJECT)
-class TestCoverageCollectorService(private val project: Project) {
+class TestStorageProcessingService(private val project: Project) {
     private val sep = File.separatorChar
 
     val id = UUID.randomUUID().toString()
@@ -41,13 +43,13 @@ class TestCoverageCollectorService(private val project: Project) {
      */
     private fun getPath(buildPath: String): String {
         // create the path for the command
-        val pluginsPath = PathManager.getPluginsPath()
-        val junitPath = "\"$pluginsPath${sep}TestSpark${sep}lib${sep}junit-4.13.jar\""
-        val mockitoPath = "\"$pluginsPath${sep}TestSpark${sep}lib${sep}mockito-core-5.0.0.jar\""
-        val hamcrestPath = "\"$pluginsPath${sep}TestSpark${sep}lib${sep}hamcrest-core-1.3.jar\""
-        val byteBuddy = "\"$pluginsPath${sep}TestSpark${sep}lib${sep}byte-buddy-1.14.6.jar\""
-        val byteBuddyAgent = "\"$pluginsPath${sep}TestSpark${sep}lib${sep}byte-buddy-agent-1.14.6.jar\""
-        return "$junitPath:$hamcrestPath:$mockitoPath:$byteBuddy:$byteBuddyAgent:$buildPath"
+        val junitPath = getLibrary("junit-4.13.jar")
+        val mockitoPath = getLibrary("mockito-core-5.0.0.jar")
+        val hamcrestPath = getLibrary("hamcrest-core-1.3.jar")
+        val byteBuddy = getLibrary("byte-buddy-1.14.6.jar")
+        val byteBuddyAgent = getLibrary("byte-buddy-agent-1.14.6.jar")
+        val sep = DataFilesUtil.classpathSeparator
+        return "$junitPath${sep}$hamcrestPath${sep}$mockitoPath${sep}$byteBuddy${sep}$byteBuddyAgent${sep}$buildPath"
     }
 
     /**
@@ -71,13 +73,19 @@ class TestCoverageCollectorService(private val project: Project) {
      */
     fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String> {
         // find the proper javac
-        val javaCompile = File(javaHomeDirectory.path).walk().filter { it.name.equals("javac") && it.isFile }.first()
+        val javaCompile = File(javaHomeDirectory.path).walk()
+            .filter {
+                val isCompilerName = if (DataFilesUtil.isWindows()) it.name.equals("javac.exe") else it.name.equals("javac")
+                isCompilerName && it.isFile
+            }
+            .first()
+
         // compile file
         val errorMsg = project.service<RunCommandLineService>().runCommandLine(
             arrayListOf(
                 javaCompile.absolutePath,
                 "-cp",
-                project.service<TestCoverageCollectorService>().getPath(projectBuildPath),
+                getPath(projectBuildPath),
                 path,
             ),
         )
@@ -92,6 +100,25 @@ class TestCoverageCollectorService(private val project: Project) {
     }
 
     /**
+     * Compiles the generated test file using the proper javac and returns a Pair
+     * indicating whether the compilation was successful and any error message encountered during compilation.
+     *
+     * @return A Pair containing a boolean indicating whether the compilation was successful
+     *         and a String containing any error message encountered during compilation.
+     */
+    fun compileTestCases(generatedTestCasesPaths: List<String>, buildPath: String, testCases: MutableList<TestCaseGeneratedByLLM>): Boolean {
+        var result = false
+        for (index in generatedTestCasesPaths.indices) {
+            val compilable = compileCode(generatedTestCasesPaths[index], buildPath).first
+            result = result || compilable
+            if (compilable) {
+                project.service<Workspace>().testGenerationData.compilableTestCases.add(testCases[index])
+            }
+        }
+        return result
+    }
+
+    /**
      * Save the generated tests to a specified directory.
      *
      * @param packageString The package string where the generated tests will be saved.
@@ -100,7 +127,7 @@ class TestCoverageCollectorService(private val project: Project) {
      * @param testFileName The name of the test file.
      * @return The path where the generated tests are saved.
      */
-    fun saveGeneratedTests(packageString: String, code: String, resultPath: String, testFileName: String): String {
+    fun saveGeneratedTest(packageString: String, code: String, resultPath: String, testFileName: String): String {
         // Generate the final path for the generated tests
         var generatedTestPath = "$resultPath${File.separatorChar}"
         packageString.split(".").forEach { directory ->
@@ -127,7 +154,7 @@ class TestCoverageCollectorService(private val project: Project) {
      * @param generatedTestPackage The package where the generated test class is located.
      * @return An empty string if the test execution is successful, otherwise an error message.
      */
-    fun createXmlFromJacoco(
+    private fun createXmlFromJacoco(
         className: String,
         dataFileName: String,
         testCaseName: String,
@@ -135,10 +162,15 @@ class TestCoverageCollectorService(private val project: Project) {
         generatedTestPackage: String,
     ): String {
         // find the proper javac
-        val javaRunner = File(javaHomeDirectory.path).walk().filter { it.name.equals("java") && it.isFile }.first()
+        val javaRunner = File(javaHomeDirectory.path).walk()
+            .filter {
+                val isJavaName = if (DataFilesUtil.isWindows()) it.name.equals("java.exe") else it.name.equals("java")
+                isJavaName && it.isFile
+            }
+            .first()
         // JaCoCo libs
-        val jacocoAgentDir = project.service<TestCoverageCollectorService>().getLibrary("jacocoagent.jar")
-        val jacocoCLIDir = project.service<TestCoverageCollectorService>().getLibrary("jacococli.jar")
+        val jacocoAgentDir = getLibrary("jacocoagent.jar")
+        val jacocoCLIDir = getLibrary("jacococli.jar")
         val sourceRoots = ModuleRootManager.getInstance(project.service<Workspace>().cutModule!!).getSourceRoots(false)
 
         // unique name
@@ -151,7 +183,7 @@ class TestCoverageCollectorService(private val project: Project) {
                 javaRunner.absolutePath,
                 "-javaagent:$jacocoAgentDir=destfile=$dataFileName.exec,append=false,includes=${project.service<Workspace>().classFQN}",
                 "-cp",
-                "${project.service<TestCoverageCollectorService>().getPath(projectBuildPath)}${project.service<TestCoverageCollectorService>().getLibrary("JUnitRunner.jar")}:$resultPath",
+                "${getPath(projectBuildPath)}${getLibrary("JUnitRunner.jar")}${DataFilesUtil.classpathSeparator}$resultPath",
                 "org.jetbrains.research.SingleJUnitTestRunner",
                 name,
             ),
@@ -196,7 +228,7 @@ class TestCoverageCollectorService(private val project: Project) {
      * @param testCaseCode The test case code.
      * @param xmlFileName The XML file name to read data from.
      */
-    fun getTestCaseFromXml(
+    private fun getTestCaseFromXml(
         testCaseId: Int,
         testCaseName: String,
         testCaseCode: String,
@@ -244,7 +276,7 @@ class TestCoverageCollectorService(private val project: Project) {
      * @param testExecutionError error output (including the thrown stack trace) during the test execution.
      * @return a set of lines that are covered in CUT during the exception happening.
      */
-    fun getExceptionData(testExecutionError: String): Pair<Boolean, Set<Int>> {
+    private fun getExceptionData(testExecutionError: String): Pair<Boolean, Set<Int>> {
         if (testExecutionError.isBlank()) {
             return Pair(false, emptySet())
         }
@@ -280,7 +312,7 @@ class TestCoverageCollectorService(private val project: Project) {
      * @param testCode new code of test
      * @param testName the name of the test
      */
-    fun updateDataWithTestCase(fileName: String, testId: Int, testName: String, testCode: String): TestCase {
+    fun processNewTestCase(fileName: String, testId: Int, testName: String, testCode: String): TestCase {
         // get buildPath
         var buildPath: String = ProjectRootManager.getInstance(project).contentRoots.first().path
         if (project.service<SettingsProjectService>().state.buildPath.isEmpty()) {
@@ -289,7 +321,7 @@ class TestCoverageCollectorService(private val project: Project) {
         }
 
         // save new test to file
-        val generatedTestPath: String = project.service<TestCoverageCollectorService>().saveGeneratedTests(
+        val generatedTestPath: String = saveGeneratedTest(
             project.service<Workspace>().testGenerationData.packageLine,
             testCode,
             project.service<Workspace>().resultPath!!,
@@ -297,14 +329,13 @@ class TestCoverageCollectorService(private val project: Project) {
         )
 
         // compilation checking
-        val compilationResult =
-            project.service<TestCoverageCollectorService>().compileCode(generatedTestPath, buildPath)
+        val compilationResult = compileCode(generatedTestPath, buildPath)
         if (!compilationResult.first) {
             project.service<TestsExecutionResultService>().addFailedTest(testId, testCode, compilationResult.second)
         } else {
             val dataFileName = "${project.service<Workspace>().resultPath!!}/jacoco-${fileName.split(".")[0]}"
 
-            val testExecutionError = project.service<TestCoverageCollectorService>().createXmlFromJacoco(
+            val testExecutionError = createXmlFromJacoco(
                 fileName.split(".")[0],
                 dataFileName,
                 testName,
@@ -315,11 +346,11 @@ class TestCoverageCollectorService(private val project: Project) {
             if (!File("$dataFileName.xml").exists()) {
                 project.service<TestsExecutionResultService>().addFailedTest(testId, testCode, testExecutionError)
             } else {
-                val testCase = project.service<TestCoverageCollectorService>().getTestCaseFromXml(
+                val testCase = getTestCaseFromXml(
                     testId,
                     testName,
                     testCode,
-                    project.service<TestCoverageCollectorService>().getExceptionData(testExecutionError).second,
+                    getExceptionData(testExecutionError).second,
                     "$dataFileName.xml",
                 )
 
@@ -329,12 +360,12 @@ class TestCoverageCollectorService(private val project: Project) {
                     project.service<TestsExecutionResultService>().addPassedTest(testId, testCode)
                 }
 
-                project.service<Workspace>().cleanFolder(project.service<Workspace>().resultPath!!)
+                DataFilesUtil.cleanFolder(project.service<Workspace>().resultPath!!)
 
                 return testCase
             }
         }
-        project.service<Workspace>().cleanFolder(project.service<Workspace>().resultPath!!)
+        DataFilesUtil.cleanFolder(project.service<Workspace>().resultPath!!)
 
         return TestCase(testId, testName, testCode, setOf(), setOf(), setOf())
     }