+++ src/main/kotlin/org/jetbrains/research/testspark/tools/Pipeline.kt
@@ -8,13 +8,13 @@ import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task
 import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.roots.ProjectRootManager
+import org.jetbrains.research.testspark.DataFilesUtil
 import org.jetbrains.research.testspark.TestSparkBundle
-import org.jetbrains.research.testspark.Util
 import org.jetbrains.research.testspark.actions.getSurroundingClass
 import org.jetbrains.research.testspark.data.FragmentToTestData
 import org.jetbrains.research.testspark.editor.Workspace
+import org.jetbrains.research.testspark.services.TestStorageProcessingService
 import org.jetbrains.research.testspark.services.CollectorService
-import org.jetbrains.research.testspark.services.TestCoverageCollectorService
 import org.jetbrains.research.testspark.tools.template.generation.ProcessManager
 
 /**
@@ -30,20 +30,10 @@ class Pipeline(
     private val project = e.project!!
 
     init {
-        project.service<Workspace>().projectClassPath =
-            ProjectRootManager.getInstance(project).contentRoots.first().path
-
-        project.service<Workspace>().testResultDirectory =
-            project.service<TestCoverageCollectorService>().testResultDirectory
-        project.service<Workspace>().testResultName = project.service<TestCoverageCollectorService>().testResultName
-        project.service<Workspace>().resultPath = project.service<TestCoverageCollectorService>().resultPath
-
-        project.service<Workspace>().baseDir =
-            "${project.service<Workspace>().testResultDirectory}${project.service<Workspace>().testResultName}-validation"
-
-        project.service<Workspace>().vFile = e.dataContext.getData(CommonDataKeys.VIRTUAL_FILE)!!
-        project.service<Workspace>().fileUrl = project.service<Workspace>().vFile!!.presentableUrl
-        project.service<Workspace>().modificationStamp = project.service<Workspace>().vFile!!.modificationStamp
+        project.service<Workspace>().projectClassPath = ProjectRootManager.getInstance(project).contentRoots.first().path
+        project.service<Workspace>().resultPath = project.service<TestStorageProcessingService>().resultPath
+        project.service<Workspace>().baseDir = "${project.service<TestStorageProcessingService>().testResultDirectory}${project.service<TestStorageProcessingService>().testResultName}-validation"
+        project.service<Workspace>().fileUrl = e.dataContext.getData(CommonDataKeys.VIRTUAL_FILE)!!.presentableUrl
 
         project.service<Workspace>().cutPsiClass = getSurroundingClass(
             e.dataContext.getData(CommonDataKeys.PSI_FILE)!!,
@@ -53,19 +43,16 @@ class Pipeline(
             .getModuleForFile(project.service<Workspace>().cutPsiClass!!.containingFile.virtualFile)!!
 
         project.service<Workspace>().classFQN = project.service<Workspace>().cutPsiClass!!.qualifiedName!!
-        project.service<Workspace>().key = getKey(project, project.service<Workspace>().classFQN!!)
 
-        project.service<Workspace>().id = project.service<TestCoverageCollectorService>().id
-
-        Util.makeTmp()
-        Util.makeDir(project.service<Workspace>().baseDir!!)
+        DataFilesUtil.makeTmp()
+        DataFilesUtil.makeDir(project.service<Workspace>().baseDir!!)
     }
 
     /**
      * Builds the project and launches generation on a separate thread.
      */
     fun runTestGeneration(processManager: ProcessManager, codeType: FragmentToTestData) {
-        clearDataBeforeTestGeneration(project, project.service<Workspace>().testResultName!!)
+        clearDataBeforeTestGeneration(project)
 
         project.service<Workspace>().technique = processManager.getTechnique()
         project.service<Workspace>().level = codeType.type!!