+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## Unreleased
 
+## 0.1.8
+### Fixed
+- Windows compatibility issue
+### Changed
+- Improve and refactor the test execution process.
+
 ## 0.1.7
 ### Fixed
 - Minor bugs in new test generation UI